Adjusting positioning of controls in Canvas overlay to fit on any screen.
Previously all child GameObjects of the Canvas were positioned in absolute
coordinates relative to the center of the canvas. This has been changed
to use the Anchors to position the controls relative to the edges of the
display itself. These changes, coupled with setting the Canvas to scale
with the screen size equally in both width and height, allows the
controls to be positioned comfortably relative to the edges of the
display. This should allow this application to look better on any tablet
or phone.

[GUIS-34]