Uncommenting Debug.Log() statements.
Disabling the stack trace printing appears to have made performance
issues go away on its own. Putting the Debug.Log() statements back in
with the stack trace disabled still allows good performance.

Commenting them out and leaving it that way in the code is not a good
method anyway. The debug information is useful. If in the future the
debug messages end up causing problems again, the Debug.Log() calls
should be abstracted away with some some means of enabling/disabling
the calls or supporting different levels of verbosity in some way.

[GUIS-34]