Explicitly set Android Target API Level to 32.
As of the update to Unity version 2022.3.40f1 the BioBird application
was no longer able to get the emg-imu board to start streaming EMG
power. The two Android API levels immediately available for this
version of Unity are 33 and 34. The levels immediately available for
2022.3.2f1 are 31 and 32. When building for 2022.3.40f1, having the
Target API Level set to "Automatic (highest installed)" broke the
ability to fully communicate with the Wearable Biofeedback application,
whereas it worked fine when building for 2022.3.2f1. Setting the API
Level to 32 (to match the highest default installed for 2022.3.2f1)
fixes the issue.

[GUIS-35]